requirements.txt
groq
flask
python-dotenv
